+++ serviceWorker.js
@@ -19,6 +19,17 @@ const precacheResources = [
   "images/backgrounds/hatchingTurtles.svg",
 ];
 
+const savePushSubscription = async (subscription) => {
+  const res = await fetch("/save-subscription", {
+    method: "POST",
+    headers: {
+      "Accept-Content": "application/json",
+    },
+    body: JSON.stringify(subscription),
+  });
+  return await res.json();
+};
+
 // When the service worker is installing, open the cache and add the precache resources to it
 self.addEventListener("install", (event) => {
   console.log("Service worker install event!");
@@ -27,8 +38,12 @@ self.addEventListener("install", (event) => {
   );
 });
 
-self.addEventListener("activate", (event) => {
-  console.log("Service worker activate event!");
+self.addEventListener("activate", async (event) => {
+  const subscription = await self.registration.pushManager.subscribe({
+    userVisibleOnly: true,
+    applicationServerKey: "",
+  });
+  await savePushSubscription(subscription);
 });
 
 // When there's an incoming fetch request, try and respond with a precached resource, otherwise fall back to the network
@@ -50,4 +65,8 @@ self.addEventListener("fetch", (event) => {
       return fetch(event.request);
     })
   );
+});
+
+self.addEventListener("push", (event) => {
+  self.registration.sendNotification(event);
 });